+++ packages/kit/src/runtime/client/parse.js
@@ -2,13 +2,18 @@ import { exec, parse_route_id } from '../../utils/routing.js';
 
 /**
  * @param {import('types').CSRPageNodeLoader[]} nodes
- * @param {Record<string, [number[], number[], number, 1?]>} dictionary
+ * @param {typeof import('__GENERATED__/client-manifest.js').dictionary} dictionary
  * @param {Record<string, (param: string) => boolean>} matchers
  * @returns {import('types').CSRRoute[]}
  */
 export function parse(nodes, dictionary, matchers) {
-	return Object.entries(dictionary).map(([id, [errors, layouts, leaf, uses_server_data]]) => {
+	return Object.entries(dictionary).map(([id, [leaf, layouts, errors]]) => {
 		const { pattern, names, types } = parse_route_id(id);
+
+		// whether or not the route uses the server data is
+		// encoded using the ones' complement, to save space
+		const uses_server_data = leaf < 0;
+		if (uses_server_data) leaf = ~leaf;
 
 		const route = {
 			id,
@@ -17,10 +22,10 @@ export function parse(nodes, dictionary, matchers) {
 				const match = pattern.exec(path);
 				if (match) return exec(match, names, types, matchers);
 			},
-			errors: errors.map((n) => nodes[n]),
-			layouts: layouts.map((n) => nodes[n]),
+			errors: [1, ...(errors || [])].map((n) => nodes[n]),
+			layouts: [0, ...(layouts || [])].map((n) => nodes[n]),
 			leaf: nodes[leaf],
-			uses_server_data: !!uses_server_data
+			uses_server_data
 		};
 
 		// bit of a hack, but ensures that layout/error node lists are the same